requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
python-dotenv==1.0.1